Words counter
Use a hashtable to store the
data – the words are the keys, and their frequencies are the values.
The output of the program is
the complete list of all the words and their frequencies in descending
order of frequencies – when
two words have the same frequency, output them by alphabetical order.
Each line of output consists
of a word, a tab, and the frequency.